Fix potential `u16` overflow for large base progress and quality (#111)
* Move combo data into effects struct

* Use i16 instead of i8 for durability

* Use u32 instead of u16 for progress and quality

* Reduce initial size of ParetoFrontBuilder
Fixes occasional thread stack overflow

* Add convenience functions

* Fix clippy warnings

+++ raphael-sim/tests/adversarial_tests.rs
@@ -12,7 +12,7 @@ const SETTINGS: Settings = Settings {
 };
 
 /// Calculate the minimum achievable Quality across all possible Condition rolls
-fn guaranteed_quality(mut settings: Settings, actions: &[Action]) -> Result<u16, &'static str> {
+fn guaranteed_quality(mut settings: Settings, actions: &[Action]) -> Result<u32, &'static str> {
     let is_valid_mask = |mut mask: i32| {
         // a 1-bit denotes an Excellent proc
         if (mask & 1) != 0 {
@@ -31,7 +31,7 @@ fn guaranteed_quality(mut settings: Settings, actions: &[Action]) -> Result<u16,
     };
 
     settings.adversarial = false;
-    let mut min_quality = u16::MAX;
+    let mut min_quality = u32::MAX;
 
     for mask in 0..(1 << actions.len()) {
         if !is_valid_mask(mask) {
